+++ idmtools_platform_file/idmtools_platform_file/platform_operations/suite_operations.py
@@ -10,7 +10,7 @@ from logging import getLogger
 from idmtools.core import ItemType
 from idmtools.entities import Suite
 from idmtools.entities.iplatform_ops.iplatform_suite_operations import IPlatformSuiteOperations
-from idmtools_platform_file.utils import FileSuite, FileExperiment
+from idmtools_platform_file.platform_operations.utils import FileSuite, FileExperiment
 
 if TYPE_CHECKING:
     from idmtools_platform_file.file_platform import FilePlatform
